fix(members): move readableApiError out of the server-action module
The settings page failed to build with "Server Actions must be async
functions." A "use server" module may only export async functions, and
readableApiError is synchronous, so exporting it from actions.ts broke every
route that imports the module.

It now lives in lib/sdp-api-error.ts, which is a plain module. actions.ts is
back to exporting only async functions and types.

Worth noting for anything similar: typecheck, lint and the unit suite all
passed while the page was unbuildable. This constraint is enforced by the
bundler, so only loading the route surfaces it.
